hero menu renders content based on selection
