+++ components/PageLayout.tsx
@@ -34,9 +34,6 @@ const PageLayout = ({ children, subtitle }: { children: ReactNode; subtitle?: st
 			<div
 				css={(theme) => css`
 					display: grid;
-					grid-template-rows: ${theme.dimensions.navbar.height}px 1fr ${theme.dimensions.footer
-							.height}px;
-					height: 100%;
 					${theme.typography.regular}
 					color: ${theme.colors.black};
 				`}
